pyproject.toml
[project]
name = "fionaa-be"
version = "0.1.0"
description = "Application assessment backend"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.9",
    "langchain-core>=0.3.59",
    "langchain-openai",
    "langgraph>=0.4.2",
    "langsmith[pytest]>=0.3.4",
    "pandas",
    "matplotlib",
    "pytest",
    "pytest-asyncio>=0.24",
    "pytest-xdist",
    "jupyter",
    "langgraph-cli[inmem]",
    "google-api-python-client>=2.128.0",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-cloud-storage>=2.10.0",
    "google-cloud-modelarmor",
    "python-dotenv",
    "pyppeteer",
    "html2text",
    "rich",
    "docling>=2.65.0",
    "langchain-mcp-adapters",
    "langchain-google-community>=3.0.4",
    "landingai-ade>=1.4.0",
    "pymupdf>=1.26.7",
    "deepagents>=0.4.1",
    "tavily-python>=0.7.21",
    "langchain-qdrant>=1.1.0",
    "langchain-docling>=2.0.0",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "streamlit>=1.40.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary,pool]>=3.3.3",
    "langchain-postgres>=0.0.17",
    "ragas>=0.4.3",
    "rapidfuzz>=3.14.3",
    "langchain-community>=0.4.1",
    "langchain-cohere>=0.5.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
